CRC32 should use and return 32bit values

+++ src/input/r_mpeg_ts.cpp
@@ -285,11 +285,11 @@ static const uint32_t crc_table[256] = {
   0xbcb4666d, 0xb8757bda, 0xb5365d03, 0xb1f740b4
 };
 
-unsigned long
+uint32_t
 mpeg_ts_reader_c::calculate_crc32(unsigned char *data,
                                   int len) {
   register int i;
-  unsigned long crc = 0xffffffff;
+  uint32_t crc = 0xffffffff;
 
   for (i = 0; i < len; i++)
     crc = (crc << 8) ^ crc_table[((crc >> 24) ^ *data++) & 0xff];
